whatsapp-greetings-railway/requirements.txt
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
pandas==2.2.2
phonenumbers==8.13.40
